Move rules from 'preact'
