+++ build-system/eslint-rules/preact.js
@@ -1,7 +1,5 @@
 'use strict';
 
-const astUtils = require('eslint/lib/rules/utils/ast-utils');
-
 // Enforces importing a Preact namespace specifier if using JSX
 //
 // Good
@@ -61,19 +59,16 @@ module.exports = {
 
     let imported = false;
     let warned = false;
-    let staticTemplate = false;
 
     return {
       Program() {
         imported = false;
         warned = false;
-        staticTemplate = false;
       },
 
       ImportNamespaceSpecifier(node) {
         if (node.local.name === 'Preact') {
           imported = true;
-          staticTemplate = node.parent.source.value !== '#preact';
         }
       },
 
@@ -83,67 +78,6 @@ module.exports = {
 
       JSXFragment(node) {
         requirePreact(node);
-      },
-
-      JSXSpreadAttribute(node) {
-        if (!staticTemplate) {
-          return;
-        }
-
-        context.report({
-          node,
-          message: [
-            'Static JSX Templates are required to use static attribute definitions',
-            'This prevents an issue with spread attributes accidentally overriding a "safe" attribute with user-provided data.',
-          ].join('\n\t'),
-        });
-      },
-
-      JSXOpeningElement(node) {
-        if (!staticTemplate) {
-          return;
-        }
-
-        const {name} = node;
-        if (name.type === 'JSXMemberExpression') {
-          return context.report({
-            node,
-            message: [
-              'Static JSX Templates are required to use regular DOM nodes or Imported Components',
-              'This prevents an issue with `<json.type />` accidentally creating a <script> node.',
-            ].join('\n\t'),
-          });
-        }
-
-        if (name.name && /^[a-z]/.test(name.name)) {
-          return;
-        }
-
-        const variable = astUtils.getVariableByName(
-          context.getScope(),
-          name.name
-        );
-
-        if (!variable || variable.defs.length === 0) {
-          return context.report({
-            node,
-            message: `Could not find ${name.name} in the lexcial scope`,
-          });
-        }
-
-        for (const def of variable.defs) {
-          if (def.type === 'ImportBinding' || def.type === 'FunctionName') {
-            continue;
-          }
-
-          context.report({
-            node,
-            message: [
-              'Static JSX Templates are required to use regular DOM nodes or Imported Components',
-              'This prevents an issue with `<UserProvidedType />` accidentally creating a <script> node.',
-            ].join('\n\t'),
-          });
-        }
       },
     };
   },
